Mocked up the landing page and contacts functionality.
Customer data is served from a mocked up JSON response, and is used
to prototype the AngularJS controllers, views and services of the
landing page.

Modified the AngularJS HTTP interceptors to use console.log instead
of alert to log errors.

Added functionality to register new contacts, and to view and edit
the registered contacts.

Run 'yo angular:route contacts/new' to create a new Angular route,
with a view and controller.

Run 'yo angular:factory contactResource' to map the underlying
REST API resource to access contacts on the backend.

Run 'yo angular:service flash' to create the new service to store
and display message in 'flash' memory.
Run 'yo angular:controller flash' to create the controller to
update the page with the flash message. Also modified the main page
layout to display the flash message.

Run 'yo angular:route contacts/view' and
'yo angular:route contacts/edit' to create the routes with
associated views and controllers.

Added an Express based Mock backend to serve out mock responses.
We won't use ngMockE2E, because it is difficult to bring it into
the yeoman+grunt pipeline.