Updated the Keycloak adapter configuration for the backend.
Also updated the Dockerfile to use the WildFly image with Keycloak
adapter. Fixed the Dockerfile to update standalone-full.xml as we
use this profile.

Updated keycloak.json config in the adapter to use the external
auth url, as the external URL needs to be used in a multi-hop
scenario.

Turned off IPv6 in the nginx DNS resolver config. Corrected the
proxy_pass configuration.